+++ packages/tools/rollup/CHANGELOG.md
@@ -1,6 +1,25 @@
 ﻿# Changelog
 
-## [0.0.1][] — `unreleased`
+## [0.0.2][] — `unreleased`
+
+> **Intended Releases**
+>
+> — `@smotaal.io/rollup@0.0.2`
+> — `@smotaal.io/tools@0.0.5`
+
+### Added
+
+- `./helpers.cjs`
+
+  - Minimal `fallbacks` scope map in `rollup.config.json`.
+
+    > Maps `scope` to `https` urls for now.
+
+  - Minimal `load` hook in `Resolver`.
+
+    > Loads scoped fallbacks from `https` only for now.
+
+## [0.0.1][]
 
 > **Intended Releases**
 >
@@ -17,8 +36,7 @@
 
   - Minimalistic path- and scope-friendly `Resolver` helper abstractions.
 
-  <!-- - Platform-agnostic `require` helper abstractions with a second optional argument for the resolved `referrer` locator (ie with the implicit `file:` scheme or other explicit scheme where supported). -->
-
 ---
 
+[0.0.2]: https://www.npmjs.com/package/@smotaal.io/rollup/v/0.0.2
 [0.0.1]: https://www.npmjs.com/package/@smotaal.io/rollup/v/0.0.1